+++ packages/plugin-ext/src/common/plugin-protocol.ts
@@ -429,7 +429,7 @@ export interface PluginDeployerStartContext {
 export const PluginDeployer = Symbol('PluginDeployer');
 export interface PluginDeployer {
 
-    start(): void;
+    start(): Promise<void>;
 
 }
 
@@ -748,7 +748,9 @@ export interface DebuggerContribution extends PlatformSpecificAdapterContributio
     enableBreakpointsFor?: {
         languageIds: string[]
     },
-    configurationAttributes?: IJSONSchema[],
+    configurationAttributes?: {
+        [request: string]: IJSONSchema
+    },
     configurationSnippets?: IJSONSchemaSnippet[],
     variables?: ScopeMap,
     adapterExecutableCommand?: string