+++ packages/plugin-ext/src/plugin/plugin-manager.ts
@@ -206,7 +206,7 @@ export class PluginManagerExtImpl implements PluginManagerExt, PluginManager {
 
         this.envExt.setQueryParameters(params.env.queryParams);
         this.envExt.setLanguage(params.env.language);
-        this.envExt.setShell(params.env.shell);
+        this.terminalService.$setShell(params.env.shell);
         this.envExt.setUIKind(params.env.uiKind);
         this.envExt.setApplicationName(params.env.appName);
         this.envExt.setAppHost(params.env.appHost);